# swagger/swagger.yaml
consumes:
    - application/json
definitions:
    AlarmPolicyType:
        type: string
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    AlertRule:
        properties:
            Algo:
                description: 统计方式
                type: string
            AlgoDisplayName:
                description: 统计方式显示名
                type: string
            DisplayName:
                description: 监控项显示名，中文名称, e.g. cpu使用率
                type: string
            ExpressionWithChinese:
                description: 中文表达式
                type: string
            For:
                description: 持续时间
                type: string
            GroupWaitTime:
                description: 等待时间
                type: string
            Interval:
                description: 告警间隔
                type: string
            Labels:
                additionalProperties:
                    type: string
                description: 自定义标签
                type: object
            Level:
                description: 告警等级
                type: string
            MonitorName:
                description: 监控项名称，英文名称，e.g. cpu_use_percent
                type: string
            Name:
                description: 告警显示名称
                type: string
            Operator:
                description: 运算操作符
                type: string
            PromqlTpl:
                $ref: '#/definitions/PromqlTemplate'
            RuleId:
                description: 告警规则唯一id
                type: string
            Threshold:
                description: 阈值
                format: double
                type: number
            Unit:
                description: 告警策略的单位
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    Contacts:
        description: Contacts 联系人组
        properties:
            EnableEmail:
                description: 是否发送邮件
                type: boolean
            EnableSms:
                description: 是否发送短信
                type: boolean
            Id:
                description: 联系人组id
                format: int64
                type: integer
            Index:
                description: 用于标识创建顺序
                format: int64
                type: integer
            Name:
                description: 联系人组名称
                type: string
            Users:
                description: 联系人组中的用户
                items:
                    $ref: '#/definitions/User'
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    CreatePolicyRequest:
        description: CreatePolicyRequest 创建告警策略的请求数据结构
        properties:
            Comment:
                description: 策略描述
                type: string
            Creator:
                description: 告警策略的创建人
                type: string
            FormPolicy:
                $ref: '#/definitions/FormPolicy'
            Labels:
                additionalProperties:
                    type: string
                description: 策略标签
                type: object
            Limit:
                description: 每条告警规则的最大告警次数
                format: int64
                type: integer
            Name:
                description: 告警策略名称
                type: string
            NotifySetup:
                $ref: '#/definitions/NotifySetup'
            Production:
                description: 业务线名称
                type: string
            PromqlPolicy:
                $ref: '#/definitions/PromqlPolicy'
            Receivers:
                $ref: '#/definitions/Receivers'
            Type:
                $ref: '#/definitions/AlarmPolicyType'
            Updater:
                description: 告警策略的最后修改人
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    CreatePolicyResponse:
        description: CreatePolicyResponse 创建告警策略的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    DeletePolicyRequest:
        description: DeletePolicyRequest 删除告警策略的请求数据结构
        properties:
            InstanceIds:
                items:
                    type: string
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    DeletePolicyResponse:
        description: DeletePolicyResponse 删除告警策略的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    DingtalkConfig:
        description: DingtalkConfig 钉钉配置
        properties:
            Id:
                format: int64
                type: integer
            Name:
                description: 机器人名称
                type: string
            Url:
                description: 机器人webhoook地址
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    ErrResponse:
        description: ErrResponse 错误响应
        properties:
            Code:
                example: 1001
                format: int64
                type: integer
            Data:
                example: {}
                type: object
            Message:
                example: 失败
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    FormPolicy:
        description: FormPolicy 表单告警策略
        properties:
            ResourceSubType:
                description: 资源子类型
                type: string
            ResourceSubTypeName:
                description: 资源子类型的中文名称
                type: string
            ResourceType:
                description: 资源类型
                type: string
            ResourceTypeName:
                description: 资源类型的中文名称
                type: string
            Resources:
                $ref: '#/definitions/Resources'
            Rules:
                description: 告警规则详情
                items:
                    $ref: '#/definitions/AlertRule'
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    ListPolicyRequest:
        description: ListPolicyRequest 获取告警策略列表的请求数据结构
        properties:
            PageNo:
                description: 当前页码，从1开始
                format: int64
                type: integer
            PageSize:
                description: 每页数据个数，默认10个, 当值为-1时不分页
                format: int64
                type: integer
            ResourceSubType:
                description: 资源子类型
                items:
                    type: string
                type: array
            ResourceType:
                description: 资源类型
                items:
                    type: string
                type: array
            SearchKey:
                description: '模糊搜索的key，允许值: name-策略名称 creator-创建者 production-业务线 resource-关联的实例'
                type: string
            SearchValue:
                type: string
            Type:
                $ref: '#/definitions/AlarmPolicyType'
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    ListPolicyResponse:
        description: ListPolicyResponse 获取告警策略列表的响应数据
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                $ref: '#/definitions/ListPolicyResponseData'
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    ListPolicyResponseData:
        properties:
            Items:
                items:
                    $ref: '#/definitions/PolicyInfo'
                type: array
            Total:
                format: int64
                type: integer
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    NoticeUser:
        description: NoticeUser 联系人
        properties:
            Email:
                description: 邮箱
                type: string
            Index:
                description: 用于标识创建顺序
                format: int64
                type: integer
            SendEmail:
                description: 是否发送邮件
                type: boolean
                x-go-name: EnableEmail
            SendSms:
                description: 是否发送短信
                type: boolean
                x-go-name: EnableSms
            Telephone:
                description: 电话
                type: string
            UserId:
                description: 用户id
                type: string
            Username:
                description: 用户名称
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
    NotifySetup:
        description: NotifySetup 通知设置
        properties:
            DingtalkRobots:
                items:
                    $ref: '#/definitions/DingtalkConfig'
                type: array
            EnableDingtalk:
                description: 是否发送到钉钉机器人
                type: boolean
            EnableFeishu:
                description: 是否发送到飞书
                type: boolean
            EnableWecom:
                description: 是否发送到企业微信
                type: boolean
            FeishuGroup:
                description: 飞书业务群
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    PolicyInfo:
        properties:
            Comment:
                description: 策略描述
                type: string
            CreatedAt:
                format: int64
                type: integer
            Creator:
                description: 告警策略的创建人
                type: string
            Enabled:
                description: 是否启用告警策略，true-启用，false-不启用
                type: boolean
            FormPolicy:
                $ref: '#/definitions/FormPolicy'
            Id:
                format: int64
                type: integer
            InstanceId:
                description: 实例id
                type: string
            Labels:
                additionalProperties:
                    type: string
                description: 策略标签
                type: object
            Limit:
                description: 每条告警规则的最大告警次数
                format: int64
                type: integer
            Name:
                description: 告警策略名称
                type: string
            NotifySetup:
                $ref: '#/definitions/NotifySetup'
            Production:
                description: 业务线名称
                type: string
            PromqlPolicy:
                $ref: '#/definitions/PromqlPolicy'
            Receivers:
                $ref: '#/definitions/Receivers'
            Type:
                $ref: '#/definitions/AlarmPolicyType'
            UpdatedAt:
                format: int64
                type: integer
            Updater:
                description: 告警策略的最后修改人
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    PromqlPolicy:
        description: PromqlPolicy 告警策略
        properties:
            AlarmCount:
                description: 告警次数
                format: int64
                type: integer
            ExpressionWithChinese:
                description: 中文表达式
                type: string
            For:
                description: 持续时间
                type: string
            GroupWaitTime:
                description: 等待时间
                type: string
            Interval:
                description: 告警间隔
                type: string
            Level:
                description: 告警等级
                type: string
            Name:
                description: 告警显示名称
                type: string
            Promql:
                description: prometheus 查询语句
                type: string
            RuleId:
                description: 告警规则id
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
    PromqlTemplate:
        description: PromQL模板
        properties:
            Algo:
                description: 聚合方式、统计方式
                type: string
            AlgoDisplayName:
                description: 统计方式的名称
                type: string
            DisplayName:
                description: 模板显示名称, e.g. CPU使用率
                type: string
            Name:
                description: 模板英文标识
                type: string
            Promql:
                description: PromQL 模板
                type: string
            ResourceSubType:
                description: 资源子类型的英文标识, 根据子类型来查找对应的模板
                type: string
            Unit:
                description: 单位
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
    Receivers:
        properties:
            ContactsGroup:
                description: 联系人组
                items:
                    $ref: '#/definitions/Contacts'
                type: array
            NoticeUsers:
                description: 联系人
                items:
                    $ref: '#/definitions/NoticeUser'
                type: array
            SendResolved:
                description: 是否发送告警恢复通知
                type: boolean
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    Resource:
        description: Resource 表示被选择的资源(资源结构后续可能会发生变化，暂时按下面结构处理)
        properties:
            Az:
                type: string
            AzName:
                type: string
            Key:
                type: string
            Lab:
                type: string
            LabName:
                type: string
            Name:
                type: string
            Region:
                type: string
            RegionName:
                type: string
            ResourceSubTypeName:
                type: string
            ResourceTypeName:
                type: string
            Value:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
...
    Resources:
        properties:
            Count:
                description: 选择的资源实例数量
                format: int64
                type: integer
            Index:
                description: 资源选择方式, 1-按服务树选择, 2-按范围选择, 3-按对象选择
                format: int64
                type: integer
            Instances:
                description: 资源实例列表
                items:
                    $ref: '#/definitions/Resource'
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
    SwitchPolicyRequest:
        description: SwitchPolicyRequest 启动或停止告警策略的请求数据结构
        properties:
            Enable:
                description: 启动或停止告警策略, true-启动 false-停止
                type: boolean
            InstanceId:
                description: 告警策略的实例id
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    SwitchPolicyResponse:
        description: SwitchPolicyResponse 启动或停止告警策略的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    UpdatePolicyRequest:
        description: UpdatePolicyRequest 更新告警策略的请求数据结构
        properties:
            Comment:
                description: 策略描述
                type: string
            CreatedAt:
                format: int64
                type: integer
            Creator:
                description: 告警策略的创建人
                type: string
            FormPolicy:
                $ref: '#/definitions/FormPolicy'
            Id:
                format: int64
                type: integer
            InstanceId:
                description: 实例id
                type: string
            Labels:
                additionalProperties:
                    type: string
                description: 策略标签
                type: object
            Limit:
                description: 每条告警规则的最大告警次数
                format: int64
                type: integer
            Name:
                description: 告警策略名称
                type: string
            NotifySetup:
                $ref: '#/definitions/NotifySetup'
            Production:
                description: 业务线名称
                type: string
            PromqlPolicy:
                $ref: '#/definitions/PromqlPolicy'
            Receivers:
                $ref: '#/definitions/Receivers'
            Type:
                $ref: '#/definitions/AlarmPolicyType'
            UpdatedAt:
                format: int64
                type: integer
            Updater:
                description: 告警策略的最后修改人
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdatePolicyResponse:
        description: UpdatePolicyResponse 更新告警策略的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    User:
        properties:
            CreatedAt:
                description: 加入用户组的时间
                format: int64
                type: integer
            Email:
                description: 用户邮箱
                type: string
            Telephone:
                description: 用户电话
                type: string
            UserId:
                description: 用户id
                type: string
            Username:
                description: 用户名称
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
host: xxxxx.swagger.galaxy.cloud
info:
    description: 告警策略
    title: 告警策略
    version: 1.0.0
paths:
...
    /alarm/v2/policy/create:
        post:
            description: 创建告警策略
            operationId: createPolicyRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/CreatePolicyRequest'
            responses:
                "200":
                    $ref: '#/responses/createPolicyResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 创建告警策略
            tags:
                - policy
    /alarm/v2/policy/delete:
        post:
            description: 删除告警策略
            operationId: deletePolicyRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/DeletePolicyRequest'
            responses:
                "200":
                    $ref: '#/responses/deletePolicyResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 删除告警策略
            tags:
                - policy
...
    /alarm/v2/policy/list:
        post:
            description: 获取告警策略列表
            operationId: listPolicyRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/ListPolicyRequest'
            responses:
                "200":
                    $ref: '#/responses/listPolicyResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 获取告警策略列表
            tags:
                - policy
    /alarm/v2/policy/switch:
        post:
            description: 启动或停止告警策略
            operationId: switchPolicyRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/SwitchPolicyRequest'
            responses:
                "200":
                    $ref: '#/responses/switchPolicyResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 启动或停止告警策略
            tags:
                - policy
    /alarm/v2/policy/update:
        post:
            description: 编辑更新告警策略
            operationId: updatePolicyRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdatePolicyRequest'
            responses:
                "200":
                    $ref: '#/responses/updatePolicyResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 编辑更新告警策略
            tags:
                - policy
...
produces:
    - application/json
responses:
...
    createPolicyResponse:
        description: ""
        schema:
            $ref: '#/definitions/CreatePolicyResponse'
...
    deletePolicyResponse:
        description: ""
        schema:
            $ref: '#/definitions/DeletePolicyResponse'
    errResponse:
        description: 发生错误时的返回信息
        schema:
            $ref: '#/definitions/ErrResponse'
...
    listPolicyResponse:
        description: ""
        schema:
            $ref: '#/definitions/ListPolicyResponse'
...
    switchPolicyResponse:
        description: ""
        schema:
            $ref: '#/definitions/SwitchPolicyResponse'
...
    updatePolicyResponse:
        description: ""
        schema:
            $ref: '#/definitions/UpdatePolicyResponse'
...
schemes:
    - http
security:
    - api_key: []
securityDefinitions:
    api_key:
        in: header
        name: api_key
        type: apiKey
    basic:
        type: basic
swagger: "2.0"